fix: roll button focus issue
If the user were to click on the Roll button, it would get focused - then, if the user decides to press space as an alternative to the button afterwards, the rollQuestion() function gets called twice resulting in one question appearing for a brief moment and then getting swapped out for a new one. This change makes it so that the button gets unfocused after pressing the space key.