Early version of the Exam component
This commit adds an early version of the exam component, which has the functionality of selecting answers and provides ability to see the score by clicking the button on the bottom and then scrolling to the top. However, the button currently TOGGLES the boolean deciding whether the game is finished or not, for debugging purposes. The score screen isn't finished as well.